requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
pydantic==2.9.2
requests==2.32.3
pyyaml>=6.0
python-multipart>=0.0.7
google-cloud-storage>=2.0.0
torchcodec
